Fix #167

+++ openHAB/OpenHABWidget.h
@@ -68,6 +68,6 @@
 - (NSString *) labelText;
 - (NSString *) labelValue;
 - (void) sendCommand:(NSString *)command;
-- (int) mappingIndexByCommand:(NSString *)command;
+- (NSUInteger) mappingIndexByCommand:(NSString *)command;
 
 @end
